Attach cursor methods to the cursor object
placeCursor and moveCursor are no longer just functions hanging out in
the grid module.  They're now attached to the cursor object.  This also
means that the game module doesn't have direct access to cursor
placement functions, which is probably a good thing.  Instead, we're
just going to pass options directly to the grid constructor.