Add method to clear values of tiles on cursor move
This still needs some work, but it currently has support for vertical
or horizontal moves with a simple addition to get diagonal support. We
need to find a way to add a ‘visited’ class to the $el and remove the
visible text on the tile. Also, this should be easy to increment a
score when we need to.